RENAME: Board DTO들 정리
